requirements.txt
astroid==2.4.2
attrs==20.2.0
boto==2.49.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.1
pytz==2020.5
requests==2.24.0
sendgrid==6.4.8
six==1.15.0
starkbank-ecdsa==1.1.0
toml==0.10.1
twilio==6.50.1
typed-ast==1.4.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
